Removed default tracer setting in TestWorkflowRule.
